requirements.txt
Django==5.0.2
djangorestframework==3.14.0
psycopg2==2.9.9
celery[redis]==5.3.6
flower==2.0.1
django-filter==23.5
twilio==9.0.0